Use '.x' rather than 'x_' when appropriate
Library user code will use the accessors like '.x' rather than accessing
the members like '.x_' directly, so make more code do that. This also
makes the code easier to drop into a polyfill for a partial native
implementation. It also makes it easier to experiment with alternate
value representations in the polyfill.