toString, toLocaleString and valueOf on float32x4 and int32x4
This patch implements toString and toValueString on a couple SIMD
types with semantics that match the spec and the current Firefox
implementation. It also throws on calls to valueOf, which cause
an exception to be thrown on math operations, which should help
avoid programmer error and make normal usage more closely resemble
that of the SIMD spec (even if, in the SIMD spec, valueOf returns
the same primitive value).

A follow-on patch should extend these changes to other SIMD types.