Truncate before range checking in float to int conversion.
Implement the range checking semantics from #315 in the polyfill and tests.

When converting a floating point value to an integer value, truncate the
floating point value before checking if it is out of range. The effect is that
negative numbers > -1.0 are converted to 0 by Uint32x4.fromFloat32x4() rather
than throwing a RangeError.