Remove the minNum and maxNum functions.
See discussion in #341.

These two functions were added to correspond with the IEEE 754 functions of the
same name. However, since EcmaScript doesn't distinguish quiet and signaling
NaNs, it is non-trivial to achieve the exact same behavior as the IEEE
functions.

The SIMD.js min() and max() functions remain. They have the same behavior as
Math.min() and Math.max().